+++ src/fswAlgorithms/smallBodyNavigation/smallBodyNavUKF/smallBodyNavUKF.cpp
@@ -64,7 +64,7 @@ void SmallBodyNavUKF::reset(uint64_t currentSimNanos) {
     /* compute UT weights to be used in the UT */
     this->wm_sigma(0) = this->kappa / (this->kappa + this->numStates);
     this->wc_sigma(0) = this->wm_sigma(0) + 1 - pow(this->alpha, 2) + this->beta;
-    for (int i = 0; i < this->numStates; i++) {
+    for (uint64_t i = 0; i < this->numStates; i++) {
         /* Assign weigths */
         this->wm_sigma(i + 1) = 1 / (2 * (this->numStates + this->kappa));
         this->wm_sigma(numStates + i + 1) = this->wm_sigma(i + 1);
@@ -106,7 +106,7 @@ void SmallBodyNavUKF::processUT(uint64_t currentSimNanos) {
     X_sigma_k.col(0) = this->x_hat_k;
 
     /* Loop to generate remaining sigma points */
-    for (int i = 0; i < this->numStates; i++) {
+    for (uint64_t i = 0; i < this->numStates; i++) {
         /* Generate sigma points */
         X_sigma_k.col(i + 1) = this->x_hat_k - sqrt(this->numStates + this->kappa) * Psqrt_k.col(i);
         X_sigma_k.col(numStates + i + 1) = x_hat_k + sqrt(this->numStates + this->kappa) * Psqrt_k.col(i);
@@ -121,7 +121,7 @@ void SmallBodyNavUKF::processUT(uint64_t currentSimNanos) {
     x_sigma_k.setZero(this->numStates);
     x_sigma_dot_k.setZero(this->numStates);
     this->x_hat_k1_.setZero(this->numStates);
-    for (int i = 0; i < this->numSigmas; i++) {
+    for (uint64_t i = 0; i < this->numSigmas; i++) {
         /* Extract sigma point */
         x_sigma_k = X_sigma_k.col(i);
 
@@ -145,7 +145,7 @@ void SmallBodyNavUKF::processUT(uint64_t currentSimNanos) {
     Eigen::VectorXd x_sigma_dev_k1_;
     x_sigma_dev_k1_.setZero(this->numStates);
     this->P_k1_.setZero(this->numStates, this->numStates);
-    for (int i = 0; i < numSigmas; i++) {
+    for (uint64_t i = 0; i < numSigmas; i++) {
         /* Compute deviation of sigma from the mean */
         x_sigma_dev_k1_ = this->X_sigma_k1_.col(i) - this->x_hat_k1_;
 
@@ -169,7 +169,7 @@ void SmallBodyNavUKF::measurementUT() {
     this->X_sigma_k1_.col(0) = this->x_hat_k1_;
 
     /* Loop to generate remaining sigma points */
-    for (int i = 0; i < this->numStates; i++) {
+    for (uint64_t i = 0; i < this->numStates; i++) {
         /* Generate sigma points */
         this->X_sigma_k1_.col(i + 1) = this->x_hat_k1_ - sqrt(this->numStates + this->kappa) * Psqrt_k1_.col(i);
         this->X_sigma_k1_.col(this->numStates + i + 1) =
@@ -180,7 +180,7 @@ void SmallBodyNavUKF::measurementUT() {
     Eigen::VectorXd x_sigma_k1_;
     x_sigma_k1_.setZero(this->numStates);
     this->y_hat_k1_.setZero(this->numMeas);
-    for (int i = 0; i < this->numSigmas; i++) {
+    for (uint64_t i = 0; i < this->numSigmas; i++) {
         /* Extract sigma point */
         x_sigma_k1_ = this->X_sigma_k1_.col(i);
 
@@ -198,7 +198,7 @@ void SmallBodyNavUKF::measurementUT() {
     y_sigma_dev_k1_.setZero(this->numStates);
     this->R_k1_.setZero(this->numMeas, this->numMeas);
     this->H.setZero(this->numStates, this->numMeas);
-    for (int i = 0; i < this->numSigmas; i++) {
+    for (uint64_t i = 0; i < this->numSigmas; i++) {
         /* Compute deviation of measurement sigma from the mean */
         x_sigma_dev_k1_ = this->X_sigma_k1_.col(i) - this->x_hat_k1_;
         y_sigma_dev_k1_ = this->Y_sigma_k1_.col(i) - this->y_hat_k1_;
