+++ src/simulation/onboardDataHandling/storageUnit/partitionedStorageUnit.cpp
@@ -19,15 +19,13 @@
 
 #include <cstdint>
 
-
-#include "partitionedStorageUnit.h"
 #include "architecture/utilities/bskLogging.h"
-
+#include "partitionedStorageUnit.h"
 
 /*! The constructor creates a partitionedStorageUnit instance with zero stored data
  @return void;
  */
-PartitionedStorageUnit::PartitionedStorageUnit(){
+PartitionedStorageUnit::PartitionedStorageUnit() {
     this->storageCapacity = 0;
     this->storedDataSum = 0;
     return;
@@ -36,15 +34,13 @@ PartitionedStorageUnit::PartitionedStorageUnit(){
 /*! Destructor.
  @return void
  */
-PartitionedStorageUnit::~PartitionedStorageUnit(){
-    return;
-}
+PartitionedStorageUnit::~PartitionedStorageUnit() { return; }
 
 /*! Custom reset function.
  @param currentClock
  @return void
  */
-void PartitionedStorageUnit::customreset(uint64_t currentClock){
+void PartitionedStorageUnit::customreset(uint64_t currentClock) {
     if (this->storageCapacity <= 0) {
         bskLogger.bskLog(BSK_INFORMATION, "The storageCapacity variable must be set to a positive value.");
     }
@@ -55,7 +51,7 @@ void PartitionedStorageUnit::customreset(uint64_t currentClock){
  @param dataName
  @return void
  */
-void PartitionedStorageUnit::addPartition(std::string dataName){
+void PartitionedStorageUnit::addPartition(std::string dataName) {
     dataInstance tmpDataInstance;
     strncpy(tmpDataInstance.dataInstanceName, dataName.c_str(), sizeof(tmpDataInstance.dataInstanceName));
     tmpDataInstance.dataInstanceSum = 0;
@@ -68,10 +64,8 @@ void PartitionedStorageUnit::addPartition(std::string dataName){
  @param data            //Vector of data to be added to each partition in partitionNames
  @return void
  */
-void PartitionedStorageUnit::setDataBuffer(std::vector<std::string> partitionNames, std::vector<long long int> data){
-
-    for (int i = 0; i < partitionNames.size(); i++)
-    {
+void PartitionedStorageUnit::setDataBuffer(std::vector<std::string> partitionNames, std::vector<long long int> data) {
+    for (size_t i = 0; i < partitionNames.size(); i++) {
         PartitionedStorageUnit::DataStorageUnitBase::setDataBuffer(partitionNames[i], data[i]);
     }
 }
